Update README.md
Create a program to calculate the total amount due for an item purchased at a store in Northern
Virginia. The total amount due is the cost of the item plus the sales tax. A user should be able to
enter the cost of the item. The program should apply a 5 percent sales tax to the item. Then the
program should display to the user the calculated sales tax along with the total cost including the
sales tax.